upratane ale stale nefunkcne :smile:

+++ underwater_world/src/Camera.cpp
@@ -1,3 +1,4 @@
+#include <cmath>
 #include <glm/glm.hpp>
 #include "Camera.h"
 
@@ -32,8 +33,8 @@ void Camera::moveTo(const glm::vec3 &pos, const glm::vec3 &rot) {
         }
 
         glm::vec3 deltaRot;
-        deltaRot.x = (distanceZ * sin(rotation.y * -1)) - (distanceZ * sin(rot.y * -1));
-        deltaRot.z = (distanceZ * -cos(rotation.y * -1)) - (distanceZ * -cos(rot.y * -1));
+        deltaRot.x = (float) ((distanceZ * std::sin(rotation.y * -1)) - (distanceZ * std::sin(rot.y * -1)));
+        deltaRot.z = (float) ((distanceZ * -std::cos(rotation.y * -1)) - (distanceZ * -std::cos(rot.y * -1)));
 
         rotation = rot;
